refactor: 사용하지 않는 클래스 제거 / seller -> actor로 package변경
