+++ LOTTO.md
@@ -4,8 +4,8 @@
     - ~~판매자는 소비자에게 로또를 반환한다.~~
 - ~~관리자는 판매자에게 정답로또를 입력한다.~~
     - ~~관리자는 판매자에게 정답으로 구성된 로또를 건내준다.~~
-- 관리자는 판매자에게 로또 하나의 가격을 정한다.
-    - 관리자는 판매자에게 로또 하나의 가격을 건내준다.
+- ~~관리자는 판매자에게 로또 하나의 가격을 정한다.~~
+    - ~~관리자는 판매자에게 로또 하나의 가격을 건내준다.~~
 - 소비자는 판매자에게 로또의 결과를 조회한다.
     - 소비자는 판매자에게 로또를 건내준다.
