--- src/main/java/domain/admin/Administer.java
@@ -1,5 +1,0 @@
-package domain.admin;
-
-public class Administer {
-
-}
