--- src/main/java/domain/user/Customer.java
@@ -1,5 +1,0 @@
-package domain.user;
-
-public class Customer {
-
-}
