--- src/test/java/domain/CustomerTest.java
@@ -1,14 +1,0 @@
-package domain;
-
-import domain.user.Customer;
-import org.junit.jupiter.api.Test;
-
-class CustomerTest {
-
-    @Test
-
-    public void test() {
-        Customer customer = new Customer();
-
-    }
-}
