+++ src/test/java/domain/actor/LottoSellerTest.java
@@ -1,4 +1,4 @@
-package domain.seller;
+package domain.actor;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
